pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "restaurant-hype-graphrag"
version = "0.1.0"
description = "GraphRAG system for restaurant hype networks (Yelp + Neo4j + LangChain)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0",
    "networkx>=3.0",
    "python-louvain>=0.16",
    "scikit-learn>=1.3",
    "neo4j>=5.0",
    "langchain>=0.1",
    "langchain-anthropic>=0.0.6",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pipeline*", "graphrag*", "api*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
